Renaming additional files
