+++ client/entrypoints/express-checkout-settings/index.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
-import PaymentRequestsPage from './payment-request-page';
+import ExpressCheckoutPage from './express-checkout-page';
 
 const container = document.getElementById(
 	'wc-stripe-payment-request-settings-container'
 );
 
 if ( container ) {
-	createRoot( container ).render( <PaymentRequestsPage /> );
+	createRoot( container ).render( <ExpressCheckoutPage /> );
 }
